+++ soc/arm/nordic_nrf/common/CMakeLists.txt
@@ -3,3 +3,10 @@
 
 zephyr_library_sources_ifdef(CONFIG_SOC_FAMILY_NRF soc_nrf_common.S)
 zephyr_include_directories(.)
+
+if (CONFIG_BUILD_WITH_TFM)
+  zephyr_sources(soc_secure.c)
+  zephyr_library_include_directories(
+    $<TARGET_PROPERTY:tfm,TFM_BINARY_DIR>/install/interface/include
+  )
+endif()
